+++ packages/core/src/__tests__/angular.styles.test.ts
@@ -1,0 +1,12 @@
+import { componentToAngular } from '../generators/angular';
+import { runTestsForTarget } from './test-generator';
+
+describe('Angular with visuallyIgnoreHostElement = false', () => {
+  runTestsForTarget({
+    options: {
+      visuallyIgnoreHostElement: false,
+    },
+    target: 'angular',
+    generator: componentToAngular,
+  });
+});