+++ packages/docs/src/routes/playground/index.tsx
@@ -18,6 +18,10 @@ const defaultTopTab: OutputFramework = 'vue';
 const defaultBottomTab: OutputFramework = 'angular';
 const defaultInputTab = 'jsx';
 
+const formatErrorToDisplay = (error: unknown) => {
+  return String(error).split('\n')[0];
+};
+
 const decodeCode = (url: URL) => {
   const code = url.searchParams.get('code');
   return code ? lzString.decompressFromBase64(code) : defaultCode;
@@ -40,7 +44,7 @@ const useOutput1 = routeLoader$(async (requestEvent) => {
     inputTab || defaultInputTab,
   ).catch((err) => {
     console.error(err);
-    return '';
+    return formatErrorToDisplay(err);
   });
   return output;
 });
@@ -56,7 +60,7 @@ const useOutput2 = routeLoader$(async (requestEvent) => {
     inputTab || defaultInputTab,
   ).catch((err) => {
     console.error(err);
-    return '';
+    return formatErrorToDisplay(err);
   });
   return output;
 });
@@ -81,14 +85,14 @@ export default component$(() => {
   const isThrottling2 = useSignal(false);
   const throttleTimeout1 = useSignal(0);
   const throttleTimeout2 = useSignal(0);
+  const errorOne = useSignal('');
+  const errorTwo = useSignal('');
 
   useVisibleTask$(() => {
     visible.value = true;
   });
 
-  useVisibleTask$(({ track }) => {
-    track(() => code.value);
-
+  const updateUrl = $(() => {
     if (code.value === defaultCode || !code.value.trim()) {
       if (location.url.searchParams.has('code')) {
         location.url.searchParams.delete('code');
@@ -109,19 +113,33 @@ export default component$(() => {
 
   const throttledCompileOne = $(
     async (code: string, outputFramework: OutputFramework, inputSyntax: InputSyntax) => {
+      updateUrl();
       if (throttleTimeout1.value) {
         clearTimeout(throttleTimeout1.value);
       }
       if (isThrottling.value) {
         throttleTimeout1.value = setTimeout(async () => {
           isThrottling.value = true;
-          output.value = await compile(code, outputFramework, inputSyntax);
+          try {
+            output.value = await compile(code, outputFramework, inputSyntax);
+            errorOne.value = '';
+          } catch (err) {
+            errorOne.value = formatErrorToDisplay(err);
+          }
+
+          updateUrl();
           isThrottling.value = false;
         }, 100) as any;
         return;
       }
       isThrottling.value = true;
-      output.value = await compile(code, outputFramework, inputSyntax);
+      try {
+        output.value = await compile(code, outputFramework, inputSyntax);
+        errorOne.value = '';
+      } catch (err) {
+        errorOne.value = formatErrorToDisplay(err);
+      }
+
       isThrottling.value = false;
     },
   );
@@ -134,13 +152,25 @@ export default component$(() => {
       if (isThrottling2.value) {
         throttleTimeout2.value = setTimeout(async () => {
           isThrottling.value = true;
-          output2.value = await compile(code, outputFramework, inputSyntax);
+          try {
+            output2.value = await compile(code, outputFramework, inputSyntax);
+            errorTwo.value = '';
+          } catch (err) {
+            errorTwo.value = formatErrorToDisplay(err);
+          }
+
           isThrottling.value = false;
         }, 100) as any;
         return;
       }
       isThrottling2.value = true;
-      output2.value = await compile(code, outputFramework, inputSyntax);
+      try {
+        output2.value = await compile(code, outputFramework, inputSyntax);
+        errorTwo.value = '';
+      } catch (err) {
+        errorTwo.value = formatErrorToDisplay(err);
+      }
+
       isThrottling2.value = false;
     },
   );
@@ -223,7 +253,7 @@ export default component$(() => {
           )}
         </div>
       </div>
-      <div class="flex gap-4 max-md:gap-0 flex-col w-full h-[90vh] max-md:!h-[calc(45dvh-35px)] border-l border-primary border-opacity-50 max-md:border-l-0 max-md:border-t">
+      <div class="flex gap-4 max-md:gap-0 flex-col w-full h-[calc(100vh-130px)] max-md:!h-[calc(45dvh-35px)] border-l border-primary border-opacity-50 max-md:border-l-0 max-md:border-t">
         <div class="flex items-center gap-2 mx-4 max-md:m-1.5 my-2 mb-0 min-h-[50px] max-md:min-h-[40px]">
           <h3 class="ml-4 max-md:ml-2 text-lg max-md:text-base">Output</h3>
           {visible.value && (
@@ -243,12 +273,15 @@ export default component$(() => {
             class="ml-16 mt-3 opacity-10 origin-top-left max-md:scale-75 max-md:ml-4"
           />
           {visible.value && (
-            <CodeEditor
-              language={languageByFramework[outputOneFramework.value]}
-              value={output.value}
-              class="absolute inset-0 h-full w-full"
-              readOnly
-            />
+            <div class="absolute inset-0 h-full w-full overflow-hidden">
+              <CodeEditor
+                language={languageByFramework[outputOneFramework.value]}
+                value={output.value}
+                class="absolute inset-0 h-full w-full"
+                readOnly
+              />
+              <ErrorWarning errorMessage={errorOne.value} />
+            </div>
           )}
         </div>
         <div class="min-h-[50px] max-md:min-h-[40px] max-md:hidden flex items-center border-primary border-opacity-50 border-t -mt-4 pt-4">
@@ -270,15 +303,33 @@ export default component$(() => {
             class="ml-16 mt-3 opacity-10 origin-top-left max-md:scale-75 max-md:ml-4"
           />
           {visible.value && (
-            <CodeEditor
-              language={languageByFramework[outputTwoFramework.value]}
-              value={output2.value}
-              class="absolute inset-0 h-full w-full"
-              readOnly
-            />
+            <div class="absolute inset-0 h-full w-full overflow-hidden">
+              <CodeEditor
+                language={languageByFramework[outputTwoFramework.value]}
+                value={output2.value}
+                class="absolute inset-0 h-full w-full"
+                readOnly
+              />
+              <ErrorWarning errorMessage={errorTwo.value} />
+            </div>
           )}
         </div>
       </div>
+    </div>
+  );
+});
+
+const ErrorWarning = component$((props: { errorMessage: string }) => {
+  // Simple alert component
+  return (
+    <div
+      class={[
+        'bg-rose-950 bg-opacity-90 border border-rose-700 text-rose-200 px-4 py-3 w-full bottom-0 absolute left-0 transition-all delay-1000',
+        !props.errorMessage && 'pointer-events-none opacity-0 translate-y-4 !delay-0',
+      ]}
+      role="alert"
+    >
+      <span class="block sm:inline min-h-[1.5em]">{props.errorMessage} &nbsp;</span>
     </div>
   );
 });