+++ packages/starter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @builder.io/create-mitosis
 
+## 0.0.8
+
+### Patch Changes
+
+- 7955e1f: Fix: `npm run start` command on Windows.
+
+## 0.0.7
+
+### Patch Changes
+
+- 4ab3ce8: Fix: remove duplicate `.eslintrc.js` `extends` config.
+
 ## 0.0.6
 
 ### Patch Changes